changed some more styles, moved textbox down using adaptive css (i think) BUT new error. When 'toggle grid' button is clicked it does not access the FInterface object temp
