+++ public/demo/index.html
@@ -17,7 +17,7 @@
 <!DOCTYPE html>
 <html style="width:100%; height:100%; color:white;">
     <head>
-        <link rel="stylesheet" href="styles.css">
+        <link rel="stylesheet" href="../visual/styles.css">
     </head>
     <body class="backgroundParent">
         <div class = "backgroundParent">
@@ -27,6 +27,6 @@
             <button id="gridToggleBtn" class="button">Toggle Grid</button>
     </div>
     </div>
-    <script type = "module" src="interface.js"></script>
+    <script type = "module" src="entry.js"></script>
     </body>
 </html>
